# src/infrastructure/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AssistantMessageOut": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationOut"
            },
            "title": "Citations",
            "type": "array"
          },
          "confidence": {
            "enum": [
              "supported",
              "insufficient_evidence",
              "refused"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "intent": {
            "enum": [
              "institutional_faq",
              "next_milestone_query",
              "attendance_confirmation",
              "report_barrier",
              "request_callback",
              "administrative_document_question",
              "clinical_or_safety_concern",
              "unknown"
            ],
            "title": "Intent",
            "type": "string"
          },
          "messageId": {
            "format": "uuid",
            "title": "Messageid",
            "type": "string"
          },
          "needsHuman": {
            "title": "Needshuman",
            "type": "boolean"
          },
          "proposedAction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProposedActionOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "messageId",
          "intent",
          "answer",
          "confidence",
          "needsHuman"
        ],
        "title": "AssistantMessageOut",
        "type": "object"
      },
...
      "CitationOut": {
        "description": "Cita comprensible: t�tulo, versi�n y secci�n. Sin URLs permanentes.",
        "properties": {
          "chunkId": {
            "format": "uuid",
            "title": "Chunkid",
            "type": "string"
          },
          "documentTitle": {
            "title": "Documenttitle",
            "type": "string"
          },
          "documentVersion": {
            "title": "Documentversion",
            "type": "string"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          }
        },
        "required": [
          "chunkId",
          "documentTitle",
          "documentVersion"
        ],
        "title": "CitationOut",
        "type": "object"
      },
      "ConfirmActionRequest": {
        "properties": {
          "operationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operationid"
          }
        },
        "title": "ConfirmActionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MediaUploadIntentRequest": {
        "properties": {
          "durationSeconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationseconds"
          },
          "mimeType": {
            "maxLength": 120,
            "title": "Mimetype",
            "type": "string"
          },
          "modality": {
            "enum": [
              "audio",
              "image"
            ],
            "title": "Modality",
            "type": "string"
          },
          "sizeBytes": {
            "title": "Sizebytes",
            "type": "integer"
          }
        },
        "required": [
          "modality",
          "mimeType",
          "sizeBytes"
        ],
        "title": "MediaUploadIntentRequest",
        "type": "object"
      },
      "MediaUploadIntentResponse": {
        "properties": {
          "expiresInSeconds": {
            "title": "Expiresinseconds",
            "type": "integer"
          },
          "mediaId": {
            "format": "uuid",
            "title": "Mediaid",
            "type": "string"
          },
          "uploadUrl": {
            "title": "Uploadurl",
            "type": "string"
          }
        },
        "required": [
          "mediaId",
          "uploadUrl",
          "expiresInSeconds"
        ],
        "title": "MediaUploadIntentResponse",
        "type": "object"
      },
...
      "ProposedActionOut": {
        "properties": {
          "kind": {
            "enum": [
              "report_barrier",
              "confirm_attendance",
              "request_callback"
            ],
            "title": "Kind",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "summary"
        ],
        "title": "ProposedActionOut",
        "type": "object"
      },
...
      "SendMessageRequest": {
        "properties": {
          "mediaId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mediaid"
          },
          "modality": {
            "default": "text",
            "enum": [
              "text",
              "audio",
              "image"
            ],
            "title": "Modality",
            "type": "string"
          },
          "operationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operationid"
          },
          "text": {
            "default": "",
            "maxLength": 1000,
            "title": "Text",
            "type": "string"
          }
        },
        "title": "SendMessageRequest",
        "type": "object"
      },
      "SessionOut": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "patientId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patientid"
          },
          "policyVersion": {
            "title": "Policyversion",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "policyVersion",
          "createdAt"
        ],
        "title": "SessionOut",
        "type": "object"
      },
...
      "StartSessionRequest": {
        "properties": {
          "patientId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patientid"
          }
        },
        "title": "StartSessionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API del piloto conectado de Kinti (Fase 2). Datos exclusivamente sint�ticos. El sem�foro representa riesgo operativo de interrupci�n de la ruta, no gravedad cl�nica. No diagnostica, no prescribe y no realiza triaje.",
    "title": "Kinti API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/assistant/media/upload-intent": {
      "post": {
        "description": "Reserva una ruta opaca y devuelve una URL firmada de corta duraci�n.\n\nEl nombre no incluye datos de la persona: s�lo un identificador aleatorio.",
        "operationId": "upload_intent_api_v1_assistant_media_upload_intent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaUploadIntentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaUploadIntentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Intent",
        "tags": [
          "asistente"
        ]
      }
    },
    "/api/v1/assistant/messages/{message_id}/confirm-action": {
      "post": {
        "description": "Convierte una acci�n propuesta en un comando idempotente.\n\nEs el �nico punto donde una conversaci�n escribe. El modelo no llega hasta\naqu�: llega la confirmaci�n expl�cita de la persona.",
        "operationId": "confirm_action_api_v1_assistant_messages__message_id__confirm_action_post",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantMessageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Confirm Action",
        "tags": [
          "asistente"
        ]
      }
    },
    "/api/v1/assistant/sessions": {
      "post": {
        "operationId": "start_session_api_v1_assistant_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Session",
        "tags": [
          "asistente"
        ]
      }
    },
    "/api/v1/assistant/sessions/{session_id}": {
      "get": {
        "operationId": "get_session_api_v1_assistant_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Session",
        "tags": [
          "asistente"
        ]
      }
    },
    "/api/v1/assistant/sessions/{session_id}/messages": {
      "post": {
        "operationId": "send_message_api_v1_assistant_sessions__session_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantMessageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Message",
        "tags": [
          "asistente"
        ]
      }
    },
...
  }
}